fabric-servers: Do not ship broken/missing updates
Fabric supports a lot of odd versions of minecraft, mainly april fools
ones that are not easy to get from vanilla.

These may be added back in at a later date however, if you want them
please do open an issue. But many of the april fools versions as i have
understood it do not ship with a server side component.