flake: Properly follow the output schema format
Change reason: Current packaging method does not follow the output
schema standard defined by nix-flakes. Notably, it fails the section
where packages are required to point to a derivation.

> packages."<system>"."<name>" = derivation;

Currently all shipped packages are defined inside of a scoped attrset
with `recurseIntoAttrs` set.

> packages.x86_64-linux.minecraftServers.vanilla-1_19

Which might not seem problematic until we look at the top-level again.

> packages.x86_64-linux.minecraftServers = { ... }

Oh no.

Build tooling, `nix flake show`, `nix flake check`, and many similar
tools run with this rule that every package is a derivation in our
flake. This is actually quite smart, as it makes extremely easy to
efficiently query which packages you may have.[1]

But it also makes it so the current layout of our flake.nix to be quite
broken, which is not ideal. We got three solutions to solve this.

1. Use `legacyPackages` over `packages`.

This is the simplest solution, as there is no requirement to have every
package to link to a derivation. But its also legacy, and not really
meant to be used.

2. Use `default.nix`.

Another solution is to throw our hands up into the air and declare that
flakes are simply not good enough for us. We fall back to using the good
old standards of `default.nix` where no such checks exist.

3. Stop scoping packages.

This is the solution i have ended up using for this commit. As all the
shipped packages in this repo are named uniquely, there is very little
reason to scope them. You can even see it as a UX improvement, since it
is now easier to point to packages. Where we go from using:

> packages.minecraftServers.quilt-1_19

To:

> packages.quilt-1_19

Ey? Thats pretty nice. And most importantly, there are no more errors
when attempting to call commands such as `nix flake check`.

Woohoo!

...

`fabric-servers` has some bad packages it seems... *sigh*

Solving that is for another commit.

1: https://discourse.nixos.org/t/flake-questions/8741/2