fabric+quilt: fix nix-prefetch-url missing sha256 generation
Due to several URLs of minecraft containing spaces in them, it will
break the command: `nix-prefetch-url`, as space is an illigal character
for a nix-store name.

This is a relatively quick fix by replacing all spaces to underscores.
The more long-term solution would be to implement
`sanitizeDerivationName` inside of Python to properly sanitize the
naming. But this works for now.