fix(minecraft-servers): don't mandate whitelist file
Previously it was impossible to actually use a non-declarative whitelist
since the startup script would generate an empty whitelist file and link
it into place. This prevents the whitelist file from being
generated/linked if the option contains no items. This entire module
deserves a rewrite and this is a pretty hacky solution, but as a quick
fix that should be backwards/compatible, it works.