fabric: use overrides for loader versions
This is a major breaking change. Where all packages previously writen as

> fabric-1_18_2-1_14_3

will now require to be written using an override:

> (fabric-1_18_2.override { loaderVersion = "0.14.3" })

This change comes in place to help manage the size of the current
repository, together with allowing testing to happen at a reaslitic
timeframe. (Just running `nix flake check` used to take 10+ minutes
due to all the derivations!)

Before this change we had packaged well over 9500 packages total, but
now after this change we are now down to 1500 packages.